Changed some hardcoded values to CometSettings.
Exported the check signature methods on camera to another class instead of creating the same methods in two class.

Rename vars that was mispelled.

Working to migrate all the direct services calling to the array that passes all services interface so we can remove the class declared and algo setFoo & getFoo on GameContext and use getService() setService()